fix potential npe

+++ tika-gui-app/src/main/java/org/tallison/tika/app/fx/tools/BatchProcess.java
@@ -287,7 +287,9 @@ public class BatchProcess {
         }
 
         void cancel() {
-            process.destroyForcibly();
+            if (process != null) {
+                process.destroyForcibly();
+            }
         }
     }
